initial support for diskless mode

+++ scripts/install.sh
@@ -1,5 +1,5 @@
 #!/bin/sh
-ver="0.1.1"
+ver="0.1.2"
 
 if [ -z ${1} ]; then
   acg_path="/opt/acg"
@@ -7,6 +7,20 @@ else
   acg_path="${1}"
 fi
 
+# system detection
+sys=$(uname -s)
+if [[ "${sys}" != "Linux" ]]; then
+  echo "Unsupport system (${sys}). ACG can only run on Linux."
+  exit 1
+fi
+
+. /etc/os-release
+if [[ "${NAME}" != "Alpine Linux" ]]; then
+  echo "Only Alpine Linux is supported currently."
+  exit 1
+fi
+
+# ACG install path
 [ ! -d "${acg_path}" ] && mkdir -p "${acg_path}"
 
 [ -f /tmp/acg.zip ] && rm /tmp/acg.zip
@@ -19,4 +33,21 @@ sed -i "s/acg_path=\"\/opt\/acg\"/acg_path=\"${escaped_acg_path}\"/g" "${acg_pat
 sed -i "s/acg_path=\"\/opt\/acg\"/acg_path=\"${escaped_acg_path}\"/g" "${acg_path}/scripts/clash-proxy"
 sed -i "s/acg_path=\"\/opt\/acg\"/acg_path=\"${escaped_acg_path}\"/g" ${acg_path}/scripts/*.sh
 chmod +x "${acg_path}/scripts/acg.sh"
+
+# arch detection
+uarch=$(uname -m)
+case "${uarch}" in
+  x86)      arch="386"
+  ;;
+  x86_64)   arch="amd64"
+  ;;
+  aarch64)  arch="armv8"
+  ;;
+  armhf)    arch="armv6"
+  ;;
+  *)        arch="${uarch}"
+  ;;
+esac
+sed -i "s/CLASH_BIN_ARCH=.*$/CLASH_BIN_ARCH=${arch}/g" "${acg_path}/files/acg-cfg-sample"
+
 ${acg_path}/scripts/acg.sh install
